Extract target building specific tests to separate file
